Fix controller end2end integration tests
